requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
sqlalchemy==2.0.31
pydantic==2.7.4
pydantic-settings==2.3.4
python-multipart==0.0.9
langgraph==0.2.36
google-generativeai==0.7.2
twilio==9.2.3
vosk==0.3.45
dateparser==1.2.0
